Safari-609.1.11.1/LayoutTests/fast/css/parsing-stroke-linecap-expected.txt

This tests checks that all of the input values for stroke-linecap parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyleInherited("stroke-linecap: butt;") is "butt"
PASS testComputedStyleInherited("stroke-linecap: round;") is "round"
PASS testComputedStyleInherited("stroke-linecap: square;") is "square"
PASS testComputedStyle(";") is "butt"
PASS test("stroke-linecap: butt;") is "butt"
PASS test("stroke-linecap: round;") is "round"
PASS test("stroke-linecap: square;") is "square"
PASS test("stroke-linecap: rnd;") is ""
PASS test("stroke-linecap: but;") is ""
PASS test("stroke-linecap: 10px;") is ""
PASS test("stroke-linecap: 10%;") is ""
PASS successfullyParsed is true

TEST COMPLETE
